model size issue

+++ jobs/fedsm_prostate/app/custom/medclip/modeling_medclip.py
@@ -138,15 +138,15 @@ class MedCLIPModel(nn.Module):
         assert vision_cls in [MedCLIPVisionModel, MedCLIPVisionModelViT], 'vision_cls should be one of [MedCLIPVisionModel, MedCLIPVisionModelViT]'
 
         self.vision_model = vision_cls(checkpoint=vision_checkpoint)
-        self.vision_model = self.vision_model.half()
+        self.vision_model = self.vision_model.to(dtype=torch.float16)
         self.text_model = MedCLIPTextModel(proj_bias=False)
-        self.text_model = self.text_model.half()
+        self.text_model = self.text_model.to(dtype=torch.float16)
 
         # learnable temperature for contrastive loss
-        self.logit_scale = nn.Parameter(torch.log(torch.tensor(1/logit_scale_init_value)))
+        self.logit_scale = nn.Parameter(torch.log(torch.tensor(1/logit_scale_init_value))).to(dtype=torch.float16)
 
         if checkpoint is not None:
-            state_dict = torch.load(os.path.join(checkpoint, constants.WEIGHTS_NAME))
+            state_dict = torch.load(os.path.join(checkpoint, constants.WEIGHTS_NAME)).to(dtype=torch.float16)
             self.load_state_dict(state_dict)
             print('load model weight from:', checkpoint)
